+++ packages/scaffold-utils/src/solana/SolanaTypesUtil.ts
@@ -8,8 +8,9 @@ import type {
 } from '@solana/web3.js'
 
 import type { SendTransactionOptions } from '@solana/wallet-adapter-base'
-import type { ConnectorType } from '@web3modal/core'
 import type { CaipNetwork } from '@web3modal/common'
+import type { Connector, ConnectorType } from '@web3modal/core'
+import type { W3mFrameTypes } from '@web3modal/wallet'
 
 export type Connection = SolanaConnection
 
@@ -25,6 +26,7 @@ export type ProviderType = {
   email?: boolean
   EIP6963?: boolean
   metadata: Metadata
+  auth?: Provider['auth']
 }
 
 export interface RequestArguments {
@@ -39,6 +41,7 @@ export interface Provider extends ProviderEventEmitterMethods {
   icon?: string
   chains: Chain[]
   type: ConnectorType
+  auth?: Pick<Connector, 'email' | 'socials' | 'showWallets' | 'walletFeatures'>
 
   // Methods
   connect: () => Promise<string>
@@ -49,7 +52,6 @@ export interface Provider extends ProviderEventEmitterMethods {
     transaction: AnyTransaction,
     options?: SendOptions
   ) => Promise<TransactionSignature>
-  signAllTransactions: (transactions: AnyTransaction[]) => Promise<SolanaWeb3Transaction[]>
   sendTransaction: (
     transaction: AnyTransaction,
     connection: Connection,
@@ -79,6 +81,10 @@ export namespace ProviderEventEmitterMethods {
     disconnect: undefined
     accountsChanged: PublicKey
     chainChanged: string
+
+    auth_rpcRequest: W3mFrameTypes.RPCRequest
+    auth_rpcSuccess: W3mFrameTypes.FrameEvent
+    auth_rpcError: Error
   }
 }