+++ packages/wagmi/exports/vue.ts
@@ -4,6 +4,7 @@ import { EVMWagmiClient, type AdapterOptions } from '@web3modal/base/adapters/ev
 import { getWeb3Modal } from '@web3modal/base/utils/library/vue'
 import { ConstantsUtil } from '@web3modal/scaffold-utils'
 import type { Config } from '@wagmi/core'
+import type { Chain } from 'viem'
 
 // -- Configs -----------------------------------------------------------
 export { defaultWagmiConfig } from '@web3modal/base/adapters/evm/wagmi'
@@ -11,11 +12,11 @@ export { defaultWagmiConfig } from '@web3modal/base/adapters/evm/wagmi'
 // -- Setup -------------------------------------------------------------------
 let appkit: AppKit | undefined = undefined
 
-type WagmiAppKitOptions = Omit<AppKitOptions, 'adapters' | 'sdkType' | 'sdkVersion'> &
+export type WagmiAppKitOptions = Omit<AppKitOptions<Chain>, 'adapters' | 'sdkType' | 'sdkVersion'> &
   AdapterOptions<Config>
 
 export function createWeb3Modal(options: WagmiAppKitOptions) {
-  const wagmiAdapter = new EVMWagmiClient() as ChainAdapter
+  const wagmiAdapter = new EVMWagmiClient()
   appkit = new AppKit({
     ...options,
     adapters: [wagmiAdapter],