+++ packages/wallet/src/W3mFrameConstants.ts
@@ -140,6 +140,9 @@ export const W3mFrameRpcConstants = {
     'personal_sign',
     'eth_signTypedData_v4',
     'eth_sendTransaction',
+    'solana_signMessage',
+    'solana_signTransaction',
+    'solana_signAndSendTransaction',
     'wallet_sendCalls',
     'wallet_grantPermissions'
   ],