+++ packages/wallet/src/W3mFrameTypes.ts
@@ -68,7 +68,10 @@ import {
   AppSetPreferredAccountRequest,
   FrameSetPreferredAccountResponse,
   WalletGetCapabilitiesRequest,
-  WalletGrantPermissionsRequest
+  WalletGrantPermissionsRequest,
+  RpcSolanaSignMessageRequest,
+  RpcSolanaSignTransactionRequest,
+  RpcSolanaSignAndSendTransactionRequest
 } from './W3mFrameSchema.js'
 import type { W3mFrameRpcConstants } from './W3mFrameConstants.js'
 import type { CaipNetworkId } from '@web3modal/common'
@@ -169,6 +172,9 @@ export namespace W3mFrameTypes {
     | z.infer<typeof RpcPersonalSignRequest>
     | z.infer<typeof RpcEthSignTypedDataV4>
     | z.infer<typeof RpcEthSendTransactionRequest>
+    | z.infer<typeof RpcSolanaSignMessageRequest>
+    | z.infer<typeof RpcSolanaSignTransactionRequest>
+    | z.infer<typeof RpcSolanaSignAndSendTransactionRequest>
     | z.infer<typeof WalletSendCallsRequest>
     | z.infer<typeof WalletGetCallsReceiptRequest>
     | z.infer<typeof WalletGetCapabilitiesRequest>